Decomposed layout
